- backend sync with GZDoom to pull in a few bugfixes and formatting corrections.

+++ source/common/textures/gametexture.cpp
@@ -430,7 +430,7 @@ CUSTOM_CVAR(Int, r_spriteadjust, 2, CVAR_ARCHIVE | CVAR_GLOBALCONFIG)
 
 float FTexCoordInfo::RowOffset(float rowoffset) const
 {
-	float scale = fabs(mScale.Y);
+	float scale = fabsf(mScale.Y);
 	if (scale == 1.f || mWorldPanning) return rowoffset;
 	else return rowoffset / scale;
 }
@@ -443,7 +443,7 @@ float FTexCoordInfo::RowOffset(float rowoffset) const
 
 float FTexCoordInfo::TextureOffset(float textureoffset) const
 {
-	float scale = fabs(mScale.X);
+	float scale = fabsf(mScale.X);
 	if (scale == 1.f || mWorldPanning) return textureoffset;
 	else return textureoffset / scale;
 }
@@ -458,7 +458,7 @@ float FTexCoordInfo::TextureAdjustWidth() const
 {
 	if (mWorldPanning)
 	{
-		float tscale = fabs(mTempScale.X);
+		float tscale = fabsf(mTempScale.X);
 		if (tscale == 1.f) return (float)mRenderWidth;
 		else return mWidth / fabs(tscale);
 	}
